Do not check hash for updates with same URL
Some projects use the same download URL for all versions of the package.
This causes that the hash changes and the package breaks. Ignore
checksums for these packages.

Note we are already not checking hashes for all installed software.
We currently do something similar for some Python libraries.

Some positional parameters of `VM-Install-From-Zip` have been removed to
make the checksum an optional parameter.