Upload Chocolatey logs in daily run
Upload the Chocolatey logs as artifacts in the daily run. I forgot this
on https://github.com/mandiant/VM-Packages/pull/22

I don't like the duplication in both actions, but I can't come up with a
better way to do this. We could move the log paths to a file if we find
ourselves modifying this often, but by now let's leave like it is.