Fix renamed packages not being pushed
Some packages (such as `notepadplusplus.vm` renamed in
https://github.com/mandiant/VM-Packages/pull/98) haven't been pushed to
myget because our action only checks modified nuspec files and not
renamed ones. Note that in a package change of name, the version is not
modified.

I have released a new version of Ana06/get-changed-files which includes
`added_modified_renamed` to allows us to easily fix this. This new
version also removes the node12 deprecation warnings.