daily: Fix badges with failures
We want to update the badges even if the tests fail, as we want to
render the number of failing packages.

Note that if updating the badges in README fail, we don't want to
commit. But I think using `always()` is ok as there won't be anything to
commit. It is similar with pushing, as there won't be anything to push.