Lint categories
Prevent use of custom (non defined) categories. Modifying the category a
package belongs to is a breaking change (breaks updates and
uninstalling), so we have fixed the categories and adding new categories
needs to be discussed.

Some packages don't have a category (we don't create a link in the tools
directory) and need to be ignored when linting.