Improve package updates automation
In case a package had GitHub release and dependencies updates, only the
GitHub release was updated. The ideas was to avoid updating many things
at once. But this is causing problems with packages (like CyberChef)
whose dependencies use download links that break often (like the ones
that don't include the version). In this case, any update to the package
which doesn't update the failing dependency will make the test fail.

Update and test dependencies and GitHub url independently. This way we
ensure we update as much as we can and prevent the described problem.