Replace success/failure badges by failures file
The success/failure badges didn't turned out to be as useful as
anticipated and they generate a noisy commit almost every day. Replace
the badges by a failures log file in the wiki that allows us to easily
see which packages failed. This still generates a commit in the wiki,
but it is better than in the project.