Limit failure log entries & display latest on top
GitHub only allow to keep action logs for a maximum of 90 days, so it
won't be very useful to keep a much longer table. Limit the number of
entries as this will make easier to read the file.

Display also latest runs on the top as those are the ones we are most
likely interested in and this is also the way GitHub displays action
runs.