Reinstall packages to prevent transient failures
We often get in the daily run the error:
`The remote file either doesn't exist, is unauthorized, or is forbidden for url`

Reinstall a package when the installation fails to prevent transient
failures in tests. Reinstall twice by default to avoid delaying test
runs (locally and in PRs) too much. Run it three times in the daily runs
as time is not an issue in this case and the transient failures are very
annoying in the logs.