visualstudio.vm: Add Visual Studio Community 2022
Add package for Visual Studio Community that install the community
package visualstudio2022community adding the following common
components:
- Visual Studio core editor
- Desktop development with C++
- .NET desktop development

I am not sure if there is a better way to install
visualstudio2022community with parameters than using `choco install` in
the install script. The community package
chocolatey-visualstudio.extension 1.11 includes a
`-DefaultParameterValues` parameter that would be a better solution (as
it would allow to change the parameters when installing the package),
but only a preview is available at the moment. A stable release of a
package is not allowed to have on a prerelease dependency.

Install the nasm assembler as dependency.