ida.diaphora.vm: Add diaphora
Note diaphora is a bit different than the existent packages, as doesn't
contain an executable, but a Python script that needs to be run from
IDA. We just unzip the tool files to the Tools directory. Adjust the
helper `VM-Install-From-Zip` to support this case.

I think we shouldn't add diaphora as default package for flare-vm, but I
would like to install it myself.