idr.vm: Install bases from GitHub
The knowledge bases were only available in Google Drive. But Google
drive is not meant to be used in scripts and downloading the files from
Google Drive doesn't seem to always work. The maintainer of IDR has
added the files to the GitHub repository and we can now take them from
the downloaded zip file.