installer.vm: Simplify dialog code
Simplify the dialog code using `System.Windows.MessageBox` to make it
easier to read and maintain. The only advantage of using a
`FlowLayoutPanel` is to be able to specify an icon, but the icon is so
small that I can't really recognise commando-vm logo in it (and flare-vm
doesn't have an icon). I think the simplified version also looks more
elegant.