VM-Clean-Up: Delete PS_Transcripts
We originally decided to prevent that `PS_Transcripts` is deleted in:
https://github.com/mandiant/VM-Packages/pull/747
But after installing FLARE-VM I would like to delete the files generated
by the installation in the `PS_Transcripts` directory. Remove the
protection, allowing to delete this folder. It is still possible to keep
it using the `excludedFolders` function argument.