[CI] Update packages number badge in README
Automatically update the number of packages in the badge in the README
when merging a PR. Note this is a bit noisy and requires skipping branch
protections for our bot, but it is the only way to get an updated exact
number in the README badge. We have been doing it in this way in capa
for a long time.