[CI] Use dynamic badge for package number
The static badge with the packages number in the README needs to be
updated via commits to the main branch. This is noisy in the commit
history and requires to modify the main branch protections to disable
_Require status checks to pass before merging_.  VM-Packages relies on
the linter and the windows-2022 test to push packages to MyGet. Relaxing
the protections may break this workflow. To solve this issues, replace
the static badge by a dynamic badge which uses a gist in our bot
account. Update this gist in the `sync_badge.yml` workflow every time a
PR is merged.