[VM-Install-IDA-Plugin] Add ZIP support
`VM-Install-IDA-Plugin` now in addition of supporting downloading an IDA
plugin file to the plugins directory, it also supports ZIPs containing a
plugin (and supporting files/directories). For ZIPs, we check if there
is an inner folder (this is the case for GH ZIPs) and if there is a
directory called 'plugins'.  We copy all files in this directory with
the exception of the `README` and the `LICENSE` file (often present in
GH repos).  The copied files must include `$pluginName`, which is used
for uninstallation.