Add issue template & automation for IDA plugins
Add issue template for IDA plugins & improve and keep consistent the
current issue templates. The new issue template supports the `send PR`
automation to create new packages for IDA plugin distributed in a
standard way: as a single file or ZIP containing a plugin (and
supporting files/directories) that need to be copied to the IDA plugins
directory.