[daily] Do not update wiki and badge if failure
Failures (such as a timeout error while testing the packages) are
causing the README badge to be updated wrongly. Change the conclusion status of the
package testing step to success even if packages fail, as the action is still
running as expected (we want to count the failures, so we expect
failures). This way we detect other errors early avoiding the wiki and
the badge to be updated.