[nodejs.vm] Relax git version to fix CI
GitHub actions has upgraded to `git` 2.46, causing that the installation
of `git.install` 2.45 fails in the CI.

Because of what seems to be a Chocolatey bug, the failure of the
`git.install` dependency does not make the packages `git` or `nodejs.vm`
to fail, causing unresolved package dependency constraints in
Chocolatey. This makes any other package to fail to install after the
installation of `nodejs.vm`.

The issue seem to cause a big amount of failure packages in our daily
run, causing that it takes more than 12 hours hours to run.