[regcool.vm] Restore hash installation
The new signature verification in `CM-Assert-Signature` using
`signtool.exe` does not work for RegCool. Restore the previous
installation checking the SHA256. As the tool is using a URL that does
not include the version, the hash will change with every update,
breaking the package. If the package is updated often, we will need to
remove it from the default configuration (replacing it total-registry)
or researching if it is possible to use `signtool.exe` to verify the
package.

Note that the previous package implementation being replaced here had an
important bug: `VM-Assert-Signature` was not called at all, as
`Get-ChildItem -Path "$toolDir\*.{exe,dll}"` does not match any file.