Fix version bug in update_package.py
The regex to match the current tool version in `update_github_url ` does not
include the `v`. However the function `get_latest_version` does include the `v`
producing a URL with a duplicated `v` when replacing the current by the latest
version. Remove the `v` in `get_latest_version` to fix the issue.