[npcap.vm] Fix npcap never finishing
Some times the npcap installation takes longer than the ahk script
expects. The script finishes with error code and without closing the
installation window and the package installation script waits
indefinitely for the process to finish. This prevents FLARE-VM
installation to continue. If you close it manually if fails (because of
the error return code).

To fix the issue, the following two steps are needed:
- Increase waiting time in the ahk script
- Set a maximum time to wait for the process in the package installation script