Add idapro.vm
Add idapro.vm, that installs IDA Pro 9 using an IDA Pro installer
`ida-pro_9*.exe` (and optionally a license file) in the Desktop.
Make also some changes to idafree.vm to keep both packages consistent.

There are some conflicts between idafree.vm and idapro.vm, as both use
the same tool name, which enable us to have a single shortcut (in the
Tools directory and the taskbar) and right-click menu option. As long as
idapro.vm is installed after idafree.vm (which should be the case using
the installer that uses alphabetical order), the installation is as I
would expect it:
- IDA Pro 9 is used in shortcut in the Tools directory and the taskbar
  (without taskbar duplication)
- The right-click menu option uses IDA launcher to find the highest
  version of IDA available.

I have kept IDA Pro 9 Desktop shortcut as I find it useful to drag
binaries to it.

We need to exclude this new package from testing, as we can't provide an
IDA installer.