Disable validation of categories in the linter
Comment out the line that validates the categories from the chocolatey install package.
This needs to be termporarily disabled until the new linter is ready and all the
packages have the new structure.
This is needed because the choco install package won't longer have a category,
it will read it from the nuspec instead.