Delete netcat package
The netcat.vm community package used in the install script of netcat fails to install.
FLARE-VM had netcat.vm with the default configuration until December 2024 where it was
removed because the community package had been unlisted and nmap.vm already installs
ncat that provides similar functionality. It is not expected that the community owners
maintain this package therefore we remove it.