[installer.vm] fix logging of failed packages
PR #1274 introduced the consistent creation of `failed_packages.txt` to
signal installation completion for automation. However, it also
inadvertently caused the installer to always log additional failure
information, regardless of actual package failures, because of checking
for the existence of the (now always present) file.

This commit addresses this by checking the actual count of failed
packages (`$failures.Count`). The detailed failure messages are now only
displayed when one or more packages have failed to install.