[CI] Test that version is increased
We were testing that the version is modified, but not that it is
increased. As chocolatey takes the latest version by default this often
causes that we test and use and old version, resulting in misleading
test results and bugs. Modify the linter to test that the local version
is higher than the one in MyGet, ensuring the version is increased.