[doc] Add installer type to new_package.yml
There is no type for installer as we do not support it in the automation
(yet). The current issue template documentation does not mention
installers, causing that users select the wrong type (`SINGLE_EXE`)
which makes the automation fail when trying to create the package.
Improve the documentation to mention installers explicitly until we
support this package type.