[linter] Fix version check
In the following PR I aimed to test that the version is increased,
instead we tested that the version is equal or higher than the one in
MyGet:
https://github.com/mandiant/VM-Packages/pull/1437

We need to test that the version is strictly higher for the packages
with modified files than the ones in MyGet. For this we need the logic
removed in the last change that obtains the modified files.