[bug] Install vcredist140.vm before python3
`python3.vm` installs a version of `vcredist140` higher than the one
specified in `vcredist140.vm` via the following dependencies:
`python3.10` -> `python3` -> `vcredist2015` -> `vcredist140`.

Adding `vcredist140.vm` to `python3.vm` as dependency to ensure a
compatible version is installed.

Note we do not need to request the new of `python3.vm` in any of the
packages. If `python3.vm` is not installed, the latest version will be
installed. If `python3.vm` is installed, then we have one of the
following situations:
- the problematic version of `vcredist140` is installed and installing
  the new `python3.vm` will fail
- a compatible version of `vcredist140` is installed and installing the
  tool will work without reinstalling the new version of `python3.vm`