[CI] Fix new package automation
The new package automation is not able to remove the `send PR` label.
This likely started failing because a change in the workflow permissions
in the mandiant organization.  Provide
actions-ecosystem/action-remove-labels the bot token so that the bot and
not `github-actions` removes the label. This ensures that the actions
has permissions to remove the label as the bot has write permissions for
issues.